Add Function generate dynamic page

+++ webservice/constants.go
@@ -2,19 +2,19 @@ package webservice
 
 // Widget Types
 const (
-	EditbleTable = "editable"
+	EditbleTable  = "editable"
 	TableWithForm = "formtable"
-	BasicTable = "table"
-	FormWidget = "formwidget"
+	BasicTable    = "table"
+	FormWidget    = "formwidget"
 )
 
-// widget map 
-func WidgetTypeMap() map[int]string {
-	widgetsmap := map[int]string {
-		0 : EditbleTable,
-		1 : TableWithForm,
-		2 : BasicTable,
-		3 : FormWidget,
+// widget map
+func WidgetTypeMap() map[int64]string {
+	widgetsmap := map[int64]string{
+		0: EditbleTable,
+		1: TableWithForm,
+		2: BasicTable,
+		3: FormWidget,
 	}
 	return widgetsmap
 }
